requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
sqlalchemy
psycopg2-binary
passlib==1.7.4
bcrypt==3.2.2
python-jose
python-multipart
email-validator
alembic
pyjwt
ruff